Added rm -f ${CATALINA_PID} to stop action of init
We've run into a few corner issues with this RPM package + the ${CATALINA_PID}file. Sometimes a previous action (tomcat8 stop) may have partially succeeded, but did not cleanup after itself and delete the PID file. So during a stop, added a explicitrm -f {$CATALINA_PID}` to do this when the stop action is called.